+++ Assets/Scripts/RoomScripts/BedRoom.cs
@@ -26,7 +26,7 @@ public class BedRoom : RoomScript
 		else
 		{
 			//timer = 45f * 1.25f * (1 - 0.25f * (level - 1));
-			timer = ValuesHolder.StandartInteractionTime;
+			timer = (float)ValuesHolder.StandartInteractionTime;
 		}
 		fixedBear.GetComponent<UnitScript>().StartMoveInRoom(Resources.Bed, GetWalkPoints(), this.gameObject);
 		timer = 150f;
